+++ internal/controller/cluster_controller_test.go
@@ -17,6 +17,7 @@ import (
 	"github.com/giantswarm/observability-operator/pkg/auth"
 	"github.com/giantswarm/observability-operator/pkg/bundle"
 	"github.com/giantswarm/observability-operator/pkg/common/organization"
+	"github.com/giantswarm/observability-operator/pkg/common/tenancy"
 	"github.com/giantswarm/observability-operator/pkg/config"
 	"github.com/giantswarm/observability-operator/pkg/monitoring"
 	"github.com/giantswarm/observability-operator/pkg/monitoring/alloy"
@@ -100,9 +101,12 @@ var _ = Describe("Cluster Controller", func() {
 				),
 			)
 
+			tenantRepository := tenancy.NewKubernetesRepository(k8sClient)
+
 			alloyMetricsService := alloy.Service{
 				Client:                 k8sClient,
 				OrganizationRepository: organizationRepository,
+				TenantRepository:       tenantRepository,
 				Config: config.Config{
 					Cluster: config.ClusterConfig{
 						Name:     "management-cluster",