+++ pkg/logging/alloy/events/service.go
@@ -15,6 +15,7 @@ import (
 	"github.com/giantswarm/observability-operator/pkg/auth"
 	"github.com/giantswarm/observability-operator/pkg/common/labels"
 	"github.com/giantswarm/observability-operator/pkg/common/organization"
+	"github.com/giantswarm/observability-operator/pkg/common/tenancy"
 	"github.com/giantswarm/observability-operator/pkg/config"
 )
 
@@ -29,6 +30,7 @@ var minimumTracingSupportVersion = semver.MustParse("1.11.0")
 type Service struct {
 	Client                 client.Client
 	OrganizationRepository organization.OrganizationRepository
+	TenantRepository       tenancy.TenantRepository
 	Config                 config.Config
 	LogsAuthManager        auth.AuthManager
 	TracesAuthManager      auth.AuthManager